refactor: moves slidable action to actions lib file
